# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media Content Publishing System API
  version: 0.1.0
  description: >
    Core backend API (Express) for authentication, content management,
    AI generation proxying, and publishing to Facebook/Instagram.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
...
paths:
...
  /content:
    get:
      tags: [Content]
      summary: List content drafts for the authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: List of drafts }
    post:
      tags: [Content]
      summary: Create a new content draft
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title: { type: string }
                body: { type: string }
                platforms:
                  type: array
                  items: { type: string, enum: [facebook, instagram] }
      responses:
        "201": { description: Draft created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...